.gitignore+=*.zwc etc. Imported and variables-renamed crasis as declzsh
LICENSE+=GPLv3

The project zplugin-crasis contains the zshrc-parsing technology so it
has been decided to import the `crasis' function, which is a ZUI app,
and to do variables and functions rename from -zcrasis* prefix into
-declzsh* prefix, etc. Also imported @declzsh-process-buffer function,
also with the renames. The application starts as declzsh. Now the next
step is to tear down the ZUI-related code (upon it will be decided: Do
we want a declzsh ZUI, semigraphical app?) and to transform the declzsh
function into a command-line utility that parses zshrc.